refactor: move RoleEntity to entities package
